Switched camera to a Activity
