+++ app/src/main/java/hypr/a255bits/com/hypr/CameraFragment/CameraActivity.kt
@@ -1,0 +1,39 @@
+package hypr.a255bits.com.hypr.CameraFragment
+
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.support.v7.app.AppCompatActivity
+import android.os.Bundle
+import android.support.v4.app.NavUtils
+import android.view.MenuItem
+
+import hypr.a255bits.com.hypr.R
+import kotlinx.android.synthetic.main.activity_camera.*
+
+class CameraActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_camera)
+        setSupportActionBar(toolbar)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+    }
+
+    override fun onStart() {
+        super.onStart()
+        cameraView.start()
+    }
+
+    override fun onPause() {
+        super.onPause()
+        cameraView.stop()
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+        when(item?.itemId){
+            android.R.id.home -> NavUtils.navigateUpFromSameTask(this)
+        }
+        return super.onOptionsItemSelected(item)
+
+    }
+}
